+++ oresat_c3/services/adcs_manager.py
