+++ oresat_c3/subsystems/adcs/guidance_functions.py
@@ -1,5 +1,5 @@
 import numpy as np
-import Quaternions as quat
+import quaternion as quat
 from Basilisk.utilities import macros
 from skyfield.framelib import itrs
