+++ oresat_c3/subsystems/adcs/kalman_filter.py
@@ -1,5 +1,5 @@
 import numpy as np
-from Quaternions import quat_mult, axis_angle_to_quaternion, quat_conjugate, hemi
+from quaternion import quat_mult, axis_angle_to_quaternion, quat_conjugate, hemi
 
 class Multiplicative_Extended_Kalman_Filter():
     def __init__(self, P_star_tracker_0, sigma_star, P_b0, sigma_gyro, sigma_bias):
