+++ oresat_c3/subsystems/adcs/quaternion.py
